Fit MARS model

Wrapper function that calls RTMB to create the model and perform the numerical optimization

MARS(
  data,
  parameters,
  map = list(),
  random = NULL,
  silent = TRUE,
  control = list(),
  run_model = TRUE,
  do_sd = TRUE,
  ...
)

Arguments

data

List of data inputs, validated by check_data

parameters

List of parameters, validated by check_parameters

map

List of parameters indicated whether they are fixed and/or how they are shared. See MakeADFun.

random

Character vector indicating the parameters that are random effects.

run_model

Logical, indicates whether the model will be fitted through nlminb.

do_sd

Logical, indicates whether the standard deviations of parameters will be calculated with sdreport.

...

Other arguments to MakeADFun.