Newton-Raphson search for fishing mortality

Performs a root finding routine to find the index of F that minimizes the difference between observed catch and the value predicted by the Baranov equation.

Newton_F(
  Cobs,
  N,
  sel,
  wt,
  M,
  fleet_area,
  q_fs,
  delta = 1,
  na = dim(N)[1],
  nr = dim(N)[2],
  ns = dim(N)[3],
  nf = length(Cobs),
  Fmax = 2,
  nitF = 5L,
  trans = c("log", "logit")
)

Arguments

Cobs

Observed catch. Vector by f

N

Stock abundance at the beginning of the time step. Array [a, r, s]

sel

Selectivity. Array [a, f, s]

wt

Fishery weight at age. Array [a, f, s]

M

Instantaneous natural mortality. Units of per year [a, s]

fleet_area

Integers that specify the region r in which fleet f operates. Vector by f

q_fs

Relative catchability of stock s for fleet f. Matrix [f, s]

delta

Numeric, the duration of time in years corresponding to the observed catch, e.g., 0.25 is a quarterly time step.

na

Integer, number of age classes

nr

Integer, number of regions

ns

Integer, number of stocks

nf

Integer, number of fleets

Fmax

Numeric, the maximum Findex value

nitF

Integer, number of iterations for the Newton-Raphson routine

trans

Whether to perform the search in log or logit space

Value

A list containing:

F_afs Fishing mortality array

F_ars Fishing mortality array

F_index Index of fishing mortality. Vector by f

CB_fs Catch (biomass) matrix

CN_afs Catch (abundance) array

VB_afs Vulnerable biomass at the beginning of the time step.

penalty Penalty term returned by posfun when F_index exceeds Fmax

fn Difference between predicted and observed catch at the last iteration. Vector by f

gr Gradient of fn with respect to F_index in either log or logit space at the last iteration. Vector by f

Details

The predicted catch for fleet f is $$ C^{\textrm{pred}}_f = \sum_s \sum_a v_{a,f,s} q_{f,s} F_f \dfrac{1 - \exp(-Z_{a,s})}{Z_{a,s}} N_{a,f,s} w_{a,f,s} $$

The Newton-Raphson routine minimizes \(f(\vec{x}) = \vec{C}^{\textrm{pred}} - \vec{C}^{\textrm{obs}}\) where the vector arrow indexes fleet.

If trans = "log", \(\vec{F} = \exp(\vec{x})\).

If trans = "logit", \(\vec{F} = F_{\textrm{max}}/(1 + \exp(-\vec{x}))\).

The gradient with respect to \(\vec{x}\) is $$ f'(\vec{x}) = \sum_s \sum_a v_{a,f,s} q_{f,s} N_{a,f,s} w_{a,f,s} \left(\dfrac{\alpha\gamma}{\beta}\right)' $$

$$ \left(\dfrac{\alpha\gamma}{\beta}\right)' = \dfrac{(\alpha\gamma' + \alpha'\gamma)\beta - \alpha\gamma\beta'}{\beta^2} $$

where

\(\alpha_f = F_f\)

\(\beta_{a,s} = Z_{a,s} = M_{a,s} + \sum_f v_{a,f,s} q_{f,s} F_f\)

\(\gamma_{a,s} = 1 - \exp(-Z_{a,s})\)

\(\beta'_{a,f,s} = v_{a,f} q_{f,s} \alpha'_f\)

\(\gamma'_{a,f,s} = \exp(-Z_{a,s})\beta'_{a,f,s}\)

If trans = "log", \(\alpha'_f = \alpha_f\).

If trans = "logit", \(\alpha'_f = F_{\textrm{max}}\exp(-x_f)/(1 + \exp(-x_f))^2\).

This function solves for \(\vec{x}\) by iterating until \(f(\vec{x})\) approaches zero. In iteration \(i+1\): $$\vec{x}_{i+1} = \vec{x}_i - \dfrac{f(\vec{x}_i)}{f'(\vec{x}_i)}$$.

Author

Q. Huynh